### 2021-10-18 ImperialCollegeLondon/champ#70 @thurber
JOSS Review Feedback
Hi CHAMP team, I'm a reviewer for your JOSS submission tracked [here](https://github.com/openjournals/joss-reviews/issues/3824). I support my team in their use of HPC resources but don't manage any clusters myself, so I'll approach my review from that perspective. For brevity, I'll aggregate my feedback into this single issue, but feel free to open new issues from this ticket as appropriate if that helps your process.

First of all, nice work! I appreciate the clean simplicity and responsiveness of the interface, and believe CHAMP could be a very useful tool in supporting computational research efforts.

### Documentation
- [x] #71
- [x] The links to the example configuration are broken.

### Testing
- [x] Having a test instance of CHAMP running somewhere that potential users or curious onlookers could try out the job submission and results viewing functionality would be helpful. Alternatively, providing a docker-compose that included an OOD container with a simple job preconfigured could accomplish the same goal. I am currently attempting to set up an OOD container, so I can let you know how it goes.

### Other thoughts
- [x] #72
- [ ] This is more of a brainstorm than a request, but it would be really neat if I could run CHAMP on a system I have control of, but it could interact with and submit jobs to an HPC cluster that I do NOT have control of. In other words, I as the admin could define not only the job scripts, but also the scripts necessary to communicate with the actual HPC cluster on behalf of the user. This would be particularly helpful in the case of a team without much scripting experience that makes heavy use of a third-party computing resource.
