### 2021-10-19 8e48661
Add applying migrations to instructions for development
