### 2021-10-20 743c447
Add ability to customise menu bar style
