### 2021-10-20 758ba38
Add documentation for customising appearance
