### 2021-11-03 ImperialCollegeLondon/champ#70 @acrlakshman
Hi @cc-a while doing initial testing by following the readme without going deep into the code, I get the following error, can you provide your input

```sh
Creating test database for alias 'default'...
System check identified no issues (0 silenced).
.............Exception during job submission
Traceback (most recent call last):
  File "/home/app/main/views.py", line 75, in create_job
    job = Job.objects.create_job(
  File "/home/app/main/models/__init__.py", line 88, in create_job
    job_id = scheduler.submit(script_path, job.work_dir)
  File "/home/app/main/tests/scheduler_mock.py", line 87, in raise_scheduler_error
    raise SchedulerError(SCHEDULER_ERROR_MESSAGE)
main.scheduler.SchedulerError: error_message
..............................................................
Ran 75 tests in 2.463s

OK
Destroying test database for alias 'default'...
```

I launched a container and installed all requirements, followed by running `python manage.py test`, if I need to start server to resolve this, please let me know, I shall take a look at this tomorrow.
