### 2021-11-03 ImperialCollegeLondon/champ#70 @cc-a
Hi @acrlakshman. All tests are actually passing there and seem to be working as expected. That output is from a logged but handled exception. The log handler prints it to the terminal in addition to the testing output. The logger output could be disabled when the tests are run but the mechanism for doing this is a bit cumbersome.
