### 2021-11-03 ImperialCollegeLondon/champ#70 @acrlakshman
> Hi @acrlakshman. All tests are actually passing there and seem to be working as expected. That output is from a logged but handled exception. The log handler prints it to the terminal in addition to the testing output. The logger output could be disabled when the tests are run but the mechanism for doing this is a bit cumbersome.

Thanks for the reply @cc-a. Can you please make note of this in a new subsection in README or somewhere of your choice, either as FAQ or NOTES.

I also faced an issue while using the Dockerfile/docker-compose, I will provide those points later today.
