### 2021-11-03 ImperialCollegeLondon/champ#70 @cc-a
> Thanks for the reply @cc-a. Can you please make note of this in a new subsection in README or somewhere of your choice, either as FAQ or NOTES.

On balance I think I prefer the slightly cumbersome silencing of the logging to muddying the documentation so I'll do that.
