### 2021-11-03 #3824 @crvernon
:mega: Mid-week rally!  Looks like @cc-a @thurber and @acrlakshman are all participating on this [issue](https://github.com/ImperialCollegeLondon/champ/issues/70) and all is rolling along well.  Let's keep pushing to get this in great shape for publication!

👏 Keep up the good work!
