### 2021-11-03 ImperialCollegeLondon/champ#70 @cc-a
> I haven't had time to explore the local HPC cluster setup yet, but I do think it may be necessary to provide some testing capabilities of the HPC integration available in some form though.

I actually started looking into this as well and think I have something close to finished that should allow testing against a demo slurm cluster.
