### 2021-11-10 ImperialCollegeLondon/champ#70 @thurber
hi @cc-a, any luck with testing against a demo slurm cluster?
