+++ requirements.txt
@@ -1,16 +1,16 @@
 asgiref==3.4.1
 certifi==2021.5.30
 charset-normalizer==2.0.4
-Django==3.2.7
+django==3.2.7
 django-filter==2.4.0
 django-tables2==2.4.0
 humanize==3.11.0
 idna==3.2
 marshmallow==3.13.0
 oauthlib==3.1.1
-python-dotenv==0.19.0
+python-dotenv==0.19.2
 pytz==2021.1
-PyYAML==5.4.1
+pyyaml==5.4.1
 requests==2.26.0
 requests-oauthlib==1.3.0
 sqlparse==0.4.2
