### 2021-11-17 #3824 @crvernon
:mega: Mid-week rally!

@thurber and @acrlakshman please provide a short update in this issue thread concerning the status of your reviews.  Thanks!

@cc-a are there any blockers on your side of things?  I would like to keep our momentum up so we may keep pushing to get this in great shape for publication!
