### 2021-11-18 7be5d33
Refactor to allow importing all modules without a portal_config.yaml
