### 2021-11-29 e30f11b
Bump humanize from 3.11.0 to 3.13.0
Bumps [humanize](https://github.com/jmoiron/humanize) from 3.11.0 to 3.13.0.
- [Release notes](https://github.com/jmoiron/humanize/releases)
- [Commits](https://github.com/jmoiron/humanize/compare/3.11.0...3.13.0)

---
updated-dependencies:
- dependency-name: humanize
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
