### 2021-12-02 #3824 @thurber
hi @crvernon, there are a few issues still outstanding in ImperialCollegeLondon/champ#70, the blocking one being a way to test the champ interface against a mock cluster. I believe @cc-a was close to implementing a solution.
