### 2021-12-08 6fdb1b4
Bump django from 3.2.7 to 4.0
Bumps [django](https://github.com/django/django) from 3.2.7 to 4.0.
- [Release notes](https://github.com/django/django/releases)
- [Commits](https://github.com/django/django/compare/3.2.7...4.0)

---
updated-dependencies:
- dependency-name: django
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
