### 2021-12-15 #3824 @crvernon
:mega: Mid-week rally!

:wave: @acrlakshman please provide a short update in this issue thread concerning the status of your reviews.

:wave: @cc-a please provide an update here on the status of your responses to the reviewers thus far.

Thanks!
