### 2021-12-15 #3824 @acrlakshman
Hi @crvernon , I need to finish going through the code, on the side I am also waiting for my comments to be addressed as noted in https://github.com/ImperialCollegeLondon/champ/issues/70.
