### 2021-12-17 c546f08
Add documentation for demo cluster
