### 2021-12-17 ImperialCollegeLondon/champ#70 @cc-a
I've (finally) managed to get a working setup for the demo cluster that also plays nicely with GitHub actions. Available for preview in the [joss-review](https://github.com/ImperialCollegeLondon/champ/tree/joss-review) branch. Instructions for use have been added to the README under the development guide section.
