### 2021-12-17 #3824 @cc-a
I've managed to sort the tricky bit with the containerised cluster for testing and will address the more minor comments in short order..
