### 2021-12-18 a945c6d
Suppress logging messages when running tests
