+++ portal/settings/settings.py
@@ -181,3 +181,5 @@ JOBS_DIR = Path(os.getenv("JOBS_DIR", str(BASE_DIR / "portal_jobs")))
 PORTAL_CONFIG_PATH = os.getenv("PORTAL_CONFIG_PATH", "portal_config.yaml")
 
 VERSION = "2.0.0"
+
+TEST_RUNNER = "main.tests.TestSuiteRunner"
