### 2021-12-18 ImperialCollegeLondon/champ#70 @cc-a
Logging messages are now suppressed when running tests in the [joss-review](https://github.com/ImperialCollegeLondon/champ/tree/joss-review) branch.
