### 2021-12-22 ImperialCollegeLondon/champ#70 @thurber
@cc-a nice work getting a docker-compose environment set up for testing CHAMP! This is cool enough that I recommend adding a paragraph about it in the paper. One minor issue in the documentation for launching the demo cluster: for the final command to work for me I had to use: `docker-compose -f demo_cluster/docker-compose.demo.yaml run --service-ports ondemand` in order to actually expose the ports to my host machine. Otherwise, I'm satisfied with my portion of the review!
