### 2022-01-04 268fb1a
Bump django from 3.2.7 to 4.0.1
Bumps [django](https://github.com/django/django) from 3.2.7 to 4.0.1.
- [Release notes](https://github.com/django/django/releases)
- [Commits](https://github.com/django/django/compare/3.2.7...4.0.1)

---
updated-dependencies:
- dependency-name: django
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
