### 2022-01-06 #3824 @crvernon
:mega: Mid-week rally!

It would be good to get this one wrapped up soon @cc-a so we can free up our reviewers.  How are things looking on your side of things for finishing up the remaining issues?  It looks like @acrlakshman has several more items to check off and @thurber has approved all.

@acrlakshman could you also provide an update to where things stand on your side of things?

Thanks!
