### 2022-01-12 #3824 @acrlakshman
Things look good with the current shape, but all changes are in `joss-review`, may I know if things will be merged into `develop` branch?
