### 2022-01-13 #3824 @crvernon
@cc-a - I want to make sure I am working on what will be the branch that is targeted for release / publication.  Your default branch is `develop` which is fine.  Are you planning on bringing the `joss` branch into `develop`?

Also, you have quite a few dependabot PRs open do you mind evaluating merging these and ensuring they don't break your code?  This may help prevent potential security issues for users.
