### 2022-01-14 8d6893b
Pin Python version to 3.10 in Dockerfile
