+++ .pre-commit-config.yaml
@@ -1,16 +1,16 @@
 repos:
   - repo: https://github.com/timothycrosley/isort
-    rev: "5.5.3"
+    rev: "5.10.1"
     hooks:
       - id: isort
         additional_dependencies: [toml]
   - repo: https://github.com/psf/black
-    rev: "20.8b1"
+    rev: "21.12b0"
     hooks:
       - id: black
         exclude: "main/migrations/"
   - repo: https://gitlab.com/pycqa/flake8
-    rev: 3.8.3
+    rev: 3.9.2
     hooks:
       - id: flake8
         exclude: main/migrations/
