+++ .github/workflows/ci.yml
@@ -8,7 +8,7 @@ jobs:
       - uses: actions/setup-python@v2
         with:
           python-version: 3.9
-      - uses: pre-commit/action@v2.0.2
+      - uses: pre-commit/action@v2.0.3
   testing:
     runs-on: ubuntu-latest
     steps:
