### 2022-01-14 #3824 @cc-a
Thanks @crvernon. All dependabot PR's have been resolved and the `joss-review` branch implementing feedback has been merged into develop.

Is it an expectation that the paper.md should be part of the mainline development branch? If not I would prefer to keep it in its own separate branch.
