### 2022-01-18 #3824 @crvernon
@cc-a - the following are some comments concerning the paper specifically:

- [ ] HPC and FAIR are used as an acronym before it has been described.  For example:  "CHAMP (CHAMP is an HPC Access and Metadata Portal)..."  This could be something like "CHAMP is a high performance computing (HPC) and metadata portal which provides an easy to use workflow for FAIR (Findability, Accessibility, Interoperability, and Reusability) scientific guiding principles data generation and publication."
- [ ] Figures are not referred to in text.  See this previous JOSS publication for how to use figure referencing:  https://joss.theoj.org/papers/10.21105/joss.03221 . Please let me know if you have questions about this.
- [ ] On line 22.  `e.g.` should always be followed by a comma
- [ ] On line 40.  Open OnDemand is given an acronym here (e.g., OOD) but it has been used earlier in text.
- [ ] Need citation for NGINX
- [ ] Need acronym breakout for UIDs
- [ ] "CHAMP is written in Python (>3.7)..." should read "CHAMP is written for use in Python (>3.7)..."  and is this supposed to be >3.7 and not >=3.7?
- [ ] YAML needs acronym breakout
- [ ] Sentence "This allows flexibility to e.g. use a job..." should read "For example, this allows flexibility to use a job..."
- [ ]  API needs acronym breakout
- [ ] OAuth2 needs citation
- [ ] Sentence "...data repositories where desired, in the case of Imperial..." should read "...data repositories where desired, as in the case of Imperial..."
- [ ] Period needed after `Dr` in "Dr Matthew Harvey"
