### 2022-01-18 #3824 @crvernon
> Thanks @crvernon. All dependabot PR's have been resolved and the `joss-review` branch implementing feedback has been merged into develop.
>
> Is it an expectation that the paper.md should be part of the mainline development branch? If not I would prefer to keep it in its own separate branch.

Yes, it is fine to have the paper in a separate branch.
