### 2022-01-24 55f5913
Implement feedback from paper review
