+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: "CHAMP is an HPC Access and Metadata Portal"
+title: "CHAMP is a HPC Access and Metadata Portal"
 tags:
   - High Performance Computing
   - Python
@@ -28,33 +28,33 @@ bibliography: refs.bib
 
 # Summary
 
-CHAMP (CHAMP is an HPC Access and Metadata Portal) provides an easy to use
-workflow for FAIR data generation and publication using high performance
-computing (HPC) resources. It provides a web based interface allowing submission
-of HPC workloads and subsequent one-click publication of the results to data
-repositories such as Zenodo. Depositions support rich metadata to repositories
-that include a full implemention of the Subject property of the DataCite
-metadata schema [@datacite].
+CHAMP is a high performance computing (HPC) and metadata portal which provides
+an easy to use workflow for FAIR (Findable, Accessible, Interoperable, and
+Reusabe) data generation and publication. It provides a web based interface
+allowing submission of HPC workloads and subsequent one-click publication of the
+results to data repositories such as Zenodo. Depositions support rich metadata
+to repositories that include a full implemention of the Subject property of the
+DataCite metadata schema [@datacite].
 
 Users submit jobs simply by choosing from pre-configured software and computing
-resource specifications:
+resource specifications (see Figure 1).
 
 ![The form for the initial step in job creation](docs/images/resources.png)
 
 The user is then asked to upload the required input files for their chosen
-software and may optionally provide a job description:
+software and may optionally provide a job description (see Figure 2).
 
 ![The form for the file upload step in job creation](docs/images/job_files.png)
 
-The status of jobs is available via a dedicated view. Publication to data
-repositories is also possible once the job has completed, with the reserved
-persistent identifier (a DOI) shown associated with the job.
+The status of jobs is available via a dedicated view (see Figure 3). Publication
+to data repositories is also possible once the job has completed, with the
+reserved persistent identifier (a DOI) shown associated with the job.
 
 ![Summary table of jobs run using the portal](docs/images/job_list.png)
 
-CHAMP makes use of the Open OnDemand [@Hudak2018] framework. This allows CHAMP
-to be portable across different HPC systems and to integrate flexibly with
-institutional infrastructure (e.g. authentication mechanisms).
+CHAMP makes use of the Open OnDemand [@Hudak2018] (OOD) framework. This allows
+CHAMP to be portable across different HPC systems and to integrate flexibly with
+institutional infrastructure (e.g., authentication mechanisms).
 
 CHAMP represents over 15 years of experimentation in HPC portal design
 [@Harvey2014]. The recent 2.0 release is a ground up rewrite to facilitate its
@@ -82,21 +82,22 @@ the selected software.
 
 # Implementation and Features
 
-CHAMP is written as a Passenger App within the Open OnDemand (OOD) framework. It
-must therefore be deployed within a local OOD instance. Use of OOD provides
-numerous advantages such as portability, a strong security model and an active
-community of users. This allows the CHAMP code base to be quite minimal with
-mechanisms for authenticaton and system scheduler interaction provided by the
-parent framework. The per-user NGINX architecture of OOD ensures all relevant
-computational processes are run under individual user UIDs.
+CHAMP is written as a Passenger App within the OOD framework. It must therefore
+be deployed within a local OOD instance. Use of OOD provides numerous advantages
+such as portability, a strong security model and an active community of
+users. This allows the CHAMP code base to be quite minimal with mechanisms for
+authenticaton and system scheduler interaction provided by the parent
+framework. The per-user NGINX [@Reese2008] architecture of OOD ensures all
+relevant computational processes are run under individual user identifiers.
 
-CHAMP is written in Python (>=3.7) using Django [@django] and supports extensive
-customisation via its main YAML configuration file. Administrators may configure
[…]
