+++ refs.bib
@@ -55,3 +55,34 @@
   title = {A metadata-driven approach to data repository design},
   journal = {Journal of Cheminformatics}
 }
+
+@article{Reese2008,
+  author = {Reese, Will},
+  title = {Nginx: The High-Performance Web Server and Reverse Proxy},
+  year = {2008},
+  issue_date = {September 2008},
+  publisher = {Belltown Media},
+  address = {Houston, TX},
+  volume = {2008},
+  number = {173},
+  issn = {1075-3583},
+  abstract = {A leaner, meaner Apache.},
+  journal = {Linux J.},
+  month = {sep},
+  articleno = {2}
+}
+
+@misc{Hardt2012,
+	series =	{Request for Comments},
+	number =	6749,
+	howpublished =	{RFC 6749},
+	publisher =	{RFC Editor},
+	doi =		{10.17487/RFC6749},
+	url =		{https://www.rfc-editor.org/info/rfc6749},
+        author =	{Hardt, Dick},
+	title =		{{The OAuth 2.0 Authorization Framework}},
+	pagetotal =	76,
+	year =		2012,
+	month =		oct,
+	abstract =	{The OAuth 2.0 authorization framework enables a third-party application to obtain limited access to an HTTP service, either on behalf of a resource owner by orchestrating an approval interaction between the resource owner and the HTTP service, or by allowing the third-party application to obtain access on its own behalf. This specification replaces and obsoletes the OAuth 1.0 protocol described in RFC 5849. {[}STANDARDS-TRACK{]}},
+}
