### 2022-01-31 a97eeb8
Bump humanize from 3.13.1 to 3.14.0
Bumps [humanize](https://github.com/jmoiron/humanize) from 3.13.1 to 3.14.0.
- [Release notes](https://github.com/jmoiron/humanize/releases)
- [Commits](https://github.com/jmoiron/humanize/compare/3.13.1...3.14.0)

---
updated-dependencies:
- dependency-name: humanize
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
