+++ requirements.txt
@@ -1,15 +1,15 @@
 asgiref==3.4.1
 certifi==2021.10.8
 charset-normalizer==2.0.10
-Django==4.0.1
+django==4.0.1
 django-filter==21.1
 django-tables2==2.4.1
-humanize==3.13.1
+humanize==3.14.0
 idna==3.3
 marshmallow==3.14.1
 oauthlib==3.1.1
 python-dotenv==0.19.2
-PyYAML==6.0
+pyyaml==6.0
 requests==2.27.1
 requests-oauthlib==1.3.0
 sqlparse==0.4.2
